fixed bug, added layers DTO, changed service, controller

+++ sandbox-web/src/main/java/com/exadel/team2/sandbox/controller/EventController.java
@@ -1,8 +1,10 @@
 package com.exadel.team2.sandbox.controller;
 
+import com.exadel.team2.sandbox.dto.EventDto;
 import com.exadel.team2.sandbox.entity.EventEntity;
 import com.exadel.team2.sandbox.service.EventService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -25,19 +27,23 @@ public class EventController {
     }
 
     @PostMapping
-    public EventEntity saveEvent(@RequestBody EventEntity eventEntity) {
-        return eventService.save(eventEntity);
+    public EventEntity saveEvent(@RequestBody EventDto eventDto) {
+        return eventService.save(eventDto);
     }
 
     @PutMapping("/{id}")
-    public EventEntity updateEvent(@PathVariable Long id, @RequestBody EventEntity eventEntity) {
-        eventEntity.setEvId(id);
-        return eventService.update(eventEntity);
+    public EventEntity updateEvent(@PathVariable Long id, @RequestBody EventDto eventDto) {
+        eventDto.setEvId(id);
+        return eventService.update(id, eventDto);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteEvent(@PathVariable Long id) {
+    public ResponseEntity<?> deleteEvent(@PathVariable Long id) {
+        if (eventService.getById(id) == null) {
+            return  ResponseEntity.notFound().build();
+        }
         eventService.delete(id);
+        return ResponseEntity.ok().build();
     }
 
 }
